+++ tests/testthat/test-tidysmd.R
@@ -382,13 +382,13 @@ test_that("groups with more than two levels return correctly", {
 })
 
 
-test_that("tidy_smd() works with `dummy_variable = TRUE`", {
+test_that("tidy_smd() works with `make_dummy_vars = TRUE`", {
   .smds <- tidy_smd(
     nhefs_weights,
     c(age, race, education),
     .group = qsmk,
     .wts = w_ate,
-    dummy_variable = TRUE
+    make_dummy_vars = TRUE
   )
 
   expect_tidy_smd_tbl(.smds, .rows = 12)