+++ NEWS.md
@@ -1,5 +1,5 @@
 # tidysmd (development version)
-* Added a `dummy_variable` argument to `tidy_smd()` to transform categorical variables to dummy variables prior to calculating the SMD (#7)
+* Added a `make_dummy_vars` argument to `tidy_smd()` to transform categorical variables to dummy variables prior to calculating the SMD (#7)
 * Added `geom_love()` (#5)
 
 # tidysmd 0.1.1